+++ charts/ccd-admin-web/values.yaml
@@ -48,6 +48,7 @@ nodejs:
     ADMINWEB_ELASTIC_INDEX_URL: "http://ccd-definition-store-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/elastic-support/index"
     ADMINWEB_ELASTIC_CASE_TYPES_URL: "http://ccd-definition-store-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/elastic-support/case-types"
     ADMINWEB_GLOBAL_SEARCH_INDEX_URL: "http://ccd-definition-store-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/elastic-support/global-search/index"
+    ADMINWEB_REINDEX_TASKS_URL: "http://ccd-definition-store-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/elastic-support/reindex/tasks"
     ADMINWEB_GET_DICTIONARY_URL: "http://ts-translation-service-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/dictionary"
 
     ADMINWEB_ROLES_WHITELIST: "ccd-import,ccd-import-validate,manage-translations,load-translations"
