+++ src/main/views/menu.html
@@ -6,7 +6,7 @@
   {% endif %}
 
   {% if adminWebAuthorization.canImportDefinition %}
-  <a href="/reindex">Reindex Tasks</a>
+    <a href="/reindex">Reindex Tasks</a>
   {% endif %}
 
   {% if adminWebAuthorization.canManageUserProfile %}
